+++ atddiff/src/lib/Compare.mli
@@ -4,11 +4,14 @@
    The findings are sorted nicely.
 *)
 
+type sort_by = Location | Hash
+
 type options = {
   (* Are fields with defaults always populated in old JSON data? *)
   json_defaults_old : bool;
   (* Are fields with defaults always populated in new JSON data? *)
   json_defaults_new : bool;
+  sort_by: sort_by;
 }
 
 val asts :