+++ atddiff/src/lib/Format_text.ml
@@ -15,8 +15,7 @@ let format_loc_text opt_loc1 opt_loc2 =
         (Loc.to_string loc1)
         (Loc.to_string loc2)
 
-let format_incompatibility_text buf (x : T.full_finding) =
-  let finding = x.finding in
+let format_incompatibility_text ~with_locations buf (finding : T.finding) =
   let is_certain =
     (* TODO: more clearly distinguish Warning from Error? *)
     match finding.kind with
@@ -37,23 +36,23 @@ let format_incompatibility_text buf (x : T.full_finding) =
     | Backward, false -> "Possible backward incompatibility"
     | Both, false -> "Possible forward and backward incompatibility"
   in
+  bprintf buf "[%s] %s:\n" finding.hash dir;
+  if with_locations then
+    bprintf buf "%s:\n"
+      (format_loc_text finding.location_old finding.location_new);
   bprintf buf "\
-%s:
-%s:
 %s
 The following types are affected:%s
 "
-    dir
-    (format_loc_text finding.location_old finding.location_new)
     finding.description
-    (x.affected_types
+    (finding.affected_types
      |> List.map (fun name -> "\n  " ^ name)
      |> String.concat "")
 
-let to_string (res : T.result) : string =
+let to_string ~with_locations (res : T.result) : string =
   let buf = Buffer.create 1000 in
   List.iter (fun x ->
-    format_incompatibility_text buf x;
+    format_incompatibility_text ~with_locations buf x;
     Buffer.add_char buf '\n'
   ) res.findings;
   Buffer.contents buf